+++ core/src/main/java/com/checku/core/user/infrastructure/UserRepositoryImpl.java
@@ -1,0 +1,18 @@
+package com.checku.core.user.infrastructure;
+
+import com.checku.core.user.domain.User;
+import com.checku.core.user.service.port.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class UserRepositoryImpl implements UserRepository {
+
+    private final UserJpaRepository userJpaRepository;
+
+    @Override
+    public User save(User user) {
+        return userJpaRepository.save(UserEntity.from(user)).toModel();
+    }
+}
