+++ core/src/main/java/com/checku/core/user/service/port/UserRepository.java
@@ -1,0 +1,8 @@
+package com.checku.core.user.service.port;
+
+import com.checku.core.user.domain.User;
+
+public interface UserRepository {
+
+    User save(User user);
+}
